requirements.txt
Django==5.1.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.4.0
django-extensions==3.2.3
python-decouple==3.8
celery==5.3.4
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
stripe==7.8.0
twilio==8.12.0
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7 